Assorted Windows cleanup
